Weapon fine tuning and descriptive updates
GENERAL
* All current German weapons (MP5, G3, MG3) now have a mitigation value
of 5. Perhaps this could become a "thing" for German weapons? Will look
into the P7 later, when it's added!

MP5
* Changed burst cost from 45 to 47, because I felt 45 was a tad bit too
low. The exact value is of course still open for further fine tuning,
but for now it is set to 47.
* Added a 3rd single shot setting;  130 accuracy at 55 TU.  This further
helps placing this weapon somewhere in between rifles and shotguns.
* Damage is lowered from 30 to 25, and mitigation increaased from 0 to
5.  This is to balance the MP5 against rifles such as the M16, G3 and AK
weapons.  Being able to burst 2 times at the same damage as these rifles
didn't seem completely fair, so a slight decrease in damage was put in
place. I did add a mitigation value of 5, however, to keep the damage
from getting too low once those partly armored aliens start appearing.
* Burst accuracy is back up at 45.  The above changes to damage and
mitigation made the previous lowering of burst accuracy to 35 seem too
harsh, so it's now back up to 45.
* Burst suppression is lowered from 50 to 40,  because otherwise a
double burst would suppress pretty much any enemy completely (which
doesn't really seem to be the role of this weapon). Two bursts of 40
suppression is still high, but I don't feel it's too high at this
moment.
* Updated the descriptions to go with these changes.

G3
* Increased range from 20 to 25.  To differentiate this rifle from the
M16, and the AK, I've set the range in the middle of the standard rifle
range (20) and the standard G3 precision rifle range (30), since this is
pretty much the same rifle as the precision rifle.
* Changed the TU cost for all single shot sets to be exactly the same as
the TU costs to fire the precision rifle, but having the exact same
accuracy values as the standard M16 rifle:  45 acc at 38 TU, 80 acc at
49 TU and 130 acc at 65 TU.  Burst fire is now 3x45 acc at 70 TU.
* Damage stays at 35, but mitigation is lowered from 10 to 5.  This
weapon does a lot of damage, since it's so much like the precision
rifle. A bit too strong for a bursting rifle. In order to make the G3
slightly less similar than the standard Precision Rifle, I felt a
decrease in mitigation was in place (which goes pretty well with how the
other 2 weapons also have a mitigation value of 5 at the moment).
* Lowered the weight ever so slightly, from 7 to 6.  Kind of felt too
heavy in actual game play.
* Updated the descriptions to go with these changes.

MG3
* Apparently my current MG3 values were already up to date on this
branch :) Some changes to the description to go along with these
changes, though, is getting updated.