G3 hotfix!
The G3 TU costs were 1% too high, which made it practically almost
impossible to fire off 1 burst and 1 snap shot at start of a round
(something that the M16 is very much capable of, and that I think all
rifles should be), so this commit is kind of a hotfix to remedy this.

Also updated the description in STRINGS.XML to go along with these
changes.