P7 update, G3 TU cost tuning, MP5 tuning
P7
-- As per Casaubon's suggestion, the P7 pistol now has a slightly lower
TU cost than the default pistol. 26 and 37. Made sure not to lower it to
the point of making a 4th shot possible. I also set the magazine
capacity back down to the previous 8, which seems to be the actual
number of bullets that fit in the real pistol magazines.
-- Updated weapon data in STRINGS.XML.

G3
-- Slight adjustment of TU costs. No big changes.
-- As a precation, in WEAPONS.XML I also set the damage of this weapon
to the default rifle damage. Not sure if these values are being used,
but might as well be safe.
-- Updated weapon data in STRINGS.XML.

MP5
-- Increased accuracy from 38 to 40.
-- Increased suppression from 40 to 50, to be able to give the same
amount of suppression as the shotgun.
-- Updated weapon data in STRINGS.XML.