P7 "burst"
Decided to do some experimenting with the P7. Seems like Xenonauts
already had a sound file for "pistol burst", so I checked it out and it
sounded pretty sweet. I timed the delay between each shot in the sound
file and matched this delay in the weapon data for the P7 and it fits
like a glove in-game :) Sooo, the P7 now has a "burst" fire, which
really just is 3 shots fired in a row. I think it's a pretty fun thing
for a pistol to have a burst mode, but now we have to decide on how to
balance the P7 from this point on (maybe lower accuracy on the "burst",
lower damage per hit, or something else, I dunno yet), or it might just
make the other pistols obsolete. I really like that it only has 8
bullets in the magazine, though, so it can't really burst away like the
MP5 or any rifle.

Anyway, have a look and let me know what you think about this change.